% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_range.R, R/geom_half_range.R
\name{geom_range}
\alias{geom_range}
\alias{geom_half_range}
\title{Plot genomic ranges}
\usage{
geom_range(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  ...,
  vjust = NULL,
  linejoin = "mitre",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE
)

geom_half_range(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  ...,
  range.orientation = "bottom",
  linejoin = "mitre",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE
)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link[ggplot2:aes]{aes()}} or
\code{\link[ggplot2:aes_]{aes_()}}. If specified and \code{inherit.aes = TRUE} (the
default), it is combined with the default mapping at the top level of the
plot. You must supply \code{mapping} if there is no plot mapping.}

\item{data}{The data to be displayed in this layer. There are three
options:

If \code{NULL}, the default, the data is inherited from the plot
data as specified in the call to \code{\link[ggplot2:ggplot]{ggplot()}}.

A \code{data.frame}, or other object, will override the plot
data. All objects will be fortified to produce a data frame. See
\code{\link[ggplot2:fortify]{fortify()}} for which variables will be created.

A \code{function} will be called with a single argument,
the plot data. The return value must be a \code{data.frame}, and
will be used as the layer data. A \code{function} can be created
from a \code{formula} (e.g. \code{~ head(.x, 10)}).}

\item{stat}{The statistical transformation to use on the data for this
layer, as a string.}

\item{position}{Position adjustment, either as a string, or the result of
a call to a position adjustment function.}

\item{...}{Other arguments passed on to \code{\link[ggplot2:layer]{layer()}}. These are
often aesthetics, used to set an aesthetic to a fixed value, like
\code{colour = "red"} or \code{size = 3}. They may also be parameters
to the paired geom/stat.}

\item{vjust}{horizontal and vertical justification of the grob.  Each
justification value should be a number between 0 and 1.  Defaults to 0.5
for both, centering each pixel over its data location.}

\item{linejoin}{Line join style (round, mitre, bevel).}

\item{na.rm}{If \code{FALSE}, the default, missing values are removed with
a warning. If \code{TRUE}, missing values are silently removed.}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.
It can also be a named logical vector to finely select the aesthetics to
display.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. \code{\link[ggplot2:borders]{borders()}}.}

\item{range.orientation}{\code{character} one of "top" or "bottom", specifying
where to plot the half range.}
}
\description{
\code{geom_range()} and \code{geom_half_range()} draw blocks/tiles with a width
specified by their \code{xstart} and \code{xend} position. The other required \code{aes},
\code{y}, is expected to be a \code{character} or \code{factor} (e.g. a transcript id/name).
These \code{geom}s are designed to represent genomic annotations that cover a
genomic range (e.g. exons, CDS). \code{geom_half_range()} plots takes advantage of
the vertical symmetry of transcript annotation by enabling plotting of a
"half range" on the top or bottom of a transcript structure. This can be
useful to free up plotting space for other transcript annotations.
}
\examples{

example_exons <-
    dplyr::tibble(
        start = c(100, 300, 500, 650),
        end = start + 100,
        strand = c("+", "+", "-", "-"),
        tx = c("A", "A", "B", "B")
    )

example_exons

base <-
    ggplot2::ggplot(
        example_exons,
        ggplot2::aes(
            xstart = start,
            xend = end,
            y = tx
        )
    )

base + geom_range()
base + geom_range(ggplot2::aes(fill = tx))

base + geom_half_range()
base + geom_half_range(range.orientation = "top")
}
