% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggtranscript-package.R
\docType{package}
\name{ggtranscript}
\alias{ggtranscript}
\title{\code{ggtranscript}: Visualizing transcript structure and annotation using
\code{ggplot2}}
\description{
The major goal of \code{ggtranscript} is the simplify the process of visualizing
transcript structure and annotation. \code{ggtranscript} achieves by introducing 4
geoms (\code{geom_range}, \code{geom_half_range}, \code{geom_intron} and \code{geom_junction})
and several helper functions that are designed to visualize transcript
annotation using \code{ggplot2}. As a \code{ggplot2} extension, \code{ggtranscript} inherits
\code{ggplot2}'s familiarity and flexibility; users can intuitively adjust
aesthetics and parameters as well as complement \code{ggtranscript} geoms with
existing \code{ggplot2} geoms.
}
