% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sod1_junctions}
\alias{sod1_junctions}
\title{Example junctions}
\format{
A \code{tibble::tibble()}:
\describe{
\item{seqnames}{\code{factor()} chromosome.}
\item{start}{\code{integer()} start position.}
\item{end}{\code{integer()} end position.}
\item{strand}{\code{factor()} strand.}
\item{mean_count}{\code{factor()} Average count across all GTEx liver samples.}
}
}
\source{
generated using \code{ggtranscript/data-raw/sod1_junctions.R}
}
\usage{
sod1_junctions
}
\description{
Junction co-ordinates and counts associated with the \emph{SOD1} gene.
Junctions counts originate from GTEx liver samples and are downloaded via the
Bioconductor package \code{snapcount}. Only unannotated junctions with a mean
count above 0.3 have been retained for this example.
}
\keyword{datasets}
