% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_diff.R
\name{to_diff}
\alias{to_diff}
\title{Obtain the differences between transcripts}
\usage{
to_diff(x, y, group_var = NULL)
}
\arguments{
\item{x}{\code{data.frame} containing ranges to compare against. \code{x} should
originate from a single transcript.}

\item{y}{\code{data.frame} containing ranges to compare with \code{x}. \code{y} can
originate from multiple transcripts.}

\item{group_var}{\code{character} if \code{y} contains more than 1 transcript,
\code{group_var} should specify the column that differentiates transcripts (e.g.
"transcript_id").}
}
\value{
a \code{data.frame} that details the differences of each \code{y} group to \code{x}.
}
\description{
\code{to_diff} is a helper function intended to facilitate visualizing transcript
differences. \code{to_diff} expects two sets of user inputted transcripts; 1. with
a single transcript to compare against (\code{x}) and 2. any number of transcripts
that will be compared to \code{x} (\code{y}).
}
\examples{

library(magrittr)

gba_ens_105_exons <- gba_ens_105 \%>\%
  dplyr::filter(type == "exon")

gba_ens_105_exons

# for example, let's compare other transcript to the mane
mane <- gba_ens_105_exons \%>\%
  dplyr::filter(transcript_name == "GBA-202")

single_tx <- gba_ens_105_exons \%>\%
  dplyr::filter(transcript_name \%in\% c("GBA-203"))

single_tx_diffs <- to_diff(
  x = mane,
  y = single_tx
)

single_tx_diffs

# y can also contain multiple transcripts
multi_tx <- gba_ens_105_exons \%>\%
  dplyr::filter(transcript_name \%in\% c("GBA-203", "GBA-201", "GBA-204"))

multi_tx_diffs <- to_diff(
  x = mane,
  y = multi_tx,
  group_var = "transcript_name"
)

multi_tx_diffs
}
