% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_intron.R
\name{to_intron}
\alias{to_intron}
\title{Convert exon co-ordinates to introns}
\usage{
to_intron(x, group_var = NULL)
}
\arguments{
\item{x}{\code{data.frame} containing exons co-ordinates (and possibly associated
transcript)}

\item{group_var}{\code{character} specifying the column with the
group, most often the transcript id/name.}
}
\description{
Taking a set of exons, \code{to_intron} will return the corresponding introns.
\code{to_intron} expects inputted exons to either 1. originate from a single
transcript or 2. be grouped via \code{group_var} such that each group corresponds
to a single transcript. In other words, \code{to_intron} assumes that exons (from
each group) do not overlap one another.
}
\examples{

example_exons <-
    dplyr::tibble(
        start = c(5, 10, 15, 20),
        end = c(7, 12, 17, 22),
        tx = c("A", "A", "B", "B")
    )

example_exons

to_intron(example_exons, group_var = "tx")
}
