% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_exon_number.R
\name{add_exon_number}
\alias{add_exon_number}
\title{Add exon number}
\usage{
add_exon_number(exons, group_var = NULL)
}
\arguments{
\item{exons}{\code{data.frame} containing exonic ranges, which can contain data
from from multiple transcripts.}

\item{group_var}{\code{character} if input \code{data.frame} contains more than 1
transcript, \code{group_var} should specify the column that differentiates
transcripts (e.g. "transcript_id").}
}
\description{
\code{add_exon_number} is a helper function that adds the exon number, i.e. the
order the exons are transcribed within a transcript, as a column in the
\code{exons}. Importantly, a "strand" column must be present within \code{exons}
in order to differentiate whether exon number should be calculated according
to ascending ("+") or descending ("-") genomic co-ordinates. A transcript
with an ambiguous strand i.e. "*" will be assumed to be "+".
}
\examples{
library(magrittr)

gba_ens_105_exons <- gba_ens_105 \%>\%
    dplyr::filter(
        type == "exon",
        transcript_name \%in\% paste0("GBA-20", 2:8)
    )

gba_ens_105_exons

gba_ens_105_exons \%>\%
    add_exon_number(group_var = "transcript_name")

# this can be useful to label exons with their order/number
base <- gba_ens_105_exons \%>\%
    add_exon_number(group_var = "transcript_name") \%>\%
    ggplot2::ggplot(
        ggplot2::aes(
            xstart = start,
            xend = end,
            y = transcript_name
        )
    ) +
    geom_range() +
    geom_intron(
        data = to_intron(gba_ens_105_exons, "transcript_name"),
        strand = "-",
        arrow.min.intron.length = 500
    )

base +
    ggplot2::geom_text(ggplot2::aes(
        x = (start + end) / 2, # plot label at midpoint of exon
        label = exon_number
    ),
    size = 1.5
    )

# for complex transcript structures or small exons, it can be useful to
# set nudge_y to plot exon numbers above their respective exons
base +
    ggplot2::geom_text(ggplot2::aes(
        x = (start + end) / 2, # plot label at midpoint of exon
        label = exon_number
    ),
    size = 2.5,
    nudge_y = 0.4
    )

# or use ggrepel::geom_label_repel to separate labels from exons
base +
    ggrepel::geom_label_repel(ggplot2::aes(
        x = (start + end) / 2,
        label = exon_number
    ),
    size = 2,
    min.segment.length = 0
    )
}
