% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggtranscript-package.R
\docType{package}
\name{ggtranscript}
\alias{ggtranscript}
\title{ggtranscript: Visualising transcript structure and annotation using ggplot2}
\description{
The major goal of ggtranscript is the simplify the process of visualizing
transcript structure and annotation. ggtranscript achieves by introducing 4
geoms (geom_range, geom_half_range, geom_intron and geom_junction) and
several helper functions that are designed to visualize transcript annotation
using ggplot2. As a ggplot2 extension, ggtranscript inherits ggplot2's
familiarity and flexibility; users can intuitively adjust aesthetics and
parameters as well as complement ggtranscript geoms with existing ggplot2
geoms.
}
