% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggtranscript-package.R
\docType{package}
\name{ggtranscript}
\alias{ggtranscript}
\title{\code{ggtranscript}: Visualizing transcript structure and annotation using
\code{ggplot2}}
\description{
The goal of \code{ggtranscript} is the simplify the process of visualizing
transcript structure and annotation. To achieve this, \code{ggtranscript}
introduces 5 new geoms (\code{geom_range()}, \code{geom_half_range()}, \code{geom_intron()},
\code{geom_junction()} and \code{geom_junction_label_repel()}) as well as several
helper functions. As a \code{ggplot2} extension, \code{ggtranscript} inherits
\code{ggplot2}'s familiarity and flexibility, enabling users to intuitively adjust
aesthetics, parameters, scales etc as well as complement \code{ggtranscript} geoms
with existing \code{ggplot2} geoms to create informative, publication-ready plots.
}
