% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sod1_annotation}
\alias{sod1_annotation}
\alias{pknox1_annotation}
\title{Example transcript annotation}
\format{
A \code{tibble::tibble()}:
\describe{
\item{seqnames}{\code{factor()} chromosome.}
\item{start}{\code{integer()} start position.}
\item{end}{\code{integer()} end position.}
\item{strand}{\code{factor()} strand.}
\item{type}{\code{factor()} E.g.gene, transcript, exon or CDS.}
\item{gene_name}{\code{character()} name of gene (GBA).}
\item{transcript_name}{\code{character()} name of transcript.}
\item{transcript_biotype}{\code{character()} biotype of transcript.}
}

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 112 rows and 8 columns.
}
\source{
generated using \code{ggtranscript/data-raw/sod1_pknox1_annotation.R}
}
\usage{
sod1_annotation

pknox1_annotation
}
\description{
Transcript annotation including the co-ordinates (hg38) of the genes,
transcripts, exons and CDS regions for \emph{SOD1} and \emph{PKNOX1}, which
originate from version 105 of the Ensembl reference annotation.
}
\keyword{datasets}
