% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_diff.R
\name{to_diff}
\alias{to_diff}
\title{Obtain the differences between transcript structure}
\usage{
to_diff(exons, ref_exons, group_var = NULL)
}
\arguments{
\item{exons}{\code{data.frame()} contains exons which can originate from multiple
transcripts differentiated by \code{group_var}.}

\item{ref_exons}{\code{data.frame()} contains exons that originate from a single
transcript, which \code{exons} will be compared against.}

\item{group_var}{\code{character()} if input data originates from more than 1
transcript, \code{group_var} must specify the column that differentiates
transcripts (e.g. "transcript_id").}
}
\value{
\code{data.frame()} details the differences between \code{exons} and
\code{ref_exons}.
}
\description{
\code{to_diff()} obtains the difference between \code{exons} from a set of transcripts
to a reference transcript (\code{ref_exons}). This can be useful when visualizing
the differences between transcript structure. \code{to_diff()} expects two sets of
input exons; 1. \code{exons} - exons from any number of transcripts that will be
compared to \code{ref_exons} and 2. \code{ref_exons} - exons from a single transcript
which acts as the reference to compare against.
}
\examples{

library(magrittr)

gba_ens_105_exons <- gba_ens_105 \%>\%
    dplyr::filter(type == "exon")

gba_ens_105_exons

# for example, let's compare other transcripts to the MANE-select transcript
mane <- gba_ens_105_exons \%>\%
    dplyr::filter(transcript_name == "GBA-202")

single_tx <- gba_ens_105_exons \%>\%
    dplyr::filter(transcript_name \%in\% c("GBA-203"))

single_tx_diffs <- to_diff(
    exons = single_tx,
    ref_exons = mane
)

single_tx_diffs

# exons can also contain multiple transcripts
multi_tx <- gba_ens_105_exons \%>\%
    dplyr::filter(transcript_name \%in\% c("GBA-203", "GBA-201", "GBA-204"))

multi_tx_diffs <- to_diff(
    exons = multi_tx,
    ref_exons = mane,
    group_var = "transcript_name"
)

multi_tx_diffs

# an example of visualising differences
mane \%>\%
    dplyr::bind_rows(multi_tx) \%>\%
    dplyr::mutate(
        transcript_name = transcript_name \%>\%
            factor(levels = c("GBA-202", "GBA-201", "GBA-203", "GBA-204"))
    ) \%>\%
    ggplot2::ggplot(
        ggplot2::aes(
            xstart = start,
            xend = end,
            y = transcript_name
        )
    ) +
    geom_range() +
    geom_range(
        data = multi_tx_diffs,
        ggplot2::aes(fill = diff_type),
        alpha = 0.2,
    ) +
    ggplot2::scale_y_discrete(
        labels = c("MANE Select", "GBA-201", "GBA-203", "GBA-204")
    )
}
